+++ authentik/events/middleware.py
@@ -21,7 +21,7 @@ from authentik.core.models import Group, User
 from authentik.events.models import Event, EventAction, Notification
 from authentik.events.utils import model_to_dict
 from authentik.lib.models import InternallyManagedMixin
-from authentik.lib.otel import should_ignore_exception
+from authentik.lib.tracing import should_ignore_exception
 from authentik.lib.utils.errors import exception_to_dict
 from authentik.stages.authenticator_static.models import StaticToken
