+++ authentik/lib/sync/outgoing/exceptions.py
@@ -1,6 +1,6 @@
 from json import JSONDecodeError
 
-from authentik.lib.otel import TracingIgnoredException
+from authentik.lib.tracing import TracingIgnoredException
 
 
 class BaseSyncException(TracingIgnoredException):
