+++ authentik/providers/saml/exceptions.py
@@ -1,6 +1,6 @@
 """authentik SAML IDP Exceptions"""
 
-from authentik.lib.otel import TracingIgnoredException
+from authentik.lib.tracing import TracingIgnoredException
 
 
 class CannotHandleAssertion(TracingIgnoredException):
