+++ authentik/sources/saml/exceptions.py
@@ -1,6 +1,6 @@
 """authentik saml source exceptions"""
 
-from authentik.lib.otel import TracingIgnoredException
+from authentik.lib.tracing import TracingIgnoredException
 
 
 class SAMLException(TracingIgnoredException):
